refactor: 인터페이스 들 따로 패키지에 보관
